requirements.txt
et-xmlfile==1.1.0
Jinja2==3.0.2
MarkupSafe==2.0.1
numpy==1.21.3
openpyxl==3.0.9
pandas==1.3.4
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
